requirements.txt
Flask==2.2.2
pyTelegramBotAPI==4.9.0